♻️ Refactor: 그룹 생성 페이지 로직 분리
